chg: move label into widgets
